backend/requirements.txt
fastapi
uvicorn[standard]
supabase
spacy
pytesseract
pdfplumber
python-multipart
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
alembic
requests
python-jose[cryptography]
passlib[bcrypt]
aiofiles
pathlib2
schedule
pytest
pytest-asyncio
tenacity
sentence-transformers
httpx
sentence-transformers

